Endpoints of branch /api/issues in module smartyard.api.issues
